refactor: test 패키지 수정
