+++ app/src/test/java/woowacourse/shopping/domain/CartTest.kt
@@ -1,11 +1,7 @@
-package woowacourse.shopping
+package woowacourse.shopping.domain
 
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import woowacourse.shopping.domain.Cart
-import woowacourse.shopping.domain.CartProduct
-import woowacourse.shopping.domain.CartProducts
-import woowacourse.shopping.domain.Product
 
 class CartTest {
     @Test
