+++ app/src/test/java/woowacourse/shopping/domain/ProductTest.kt
@@ -1,8 +1,7 @@
-package woowacourse.shopping
+package woowacourse.shopping.domain
 
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import woowacourse.shopping.domain.Product
 
 class ProductTest {
     @Test
